Moved initialization data back
